Final logic touches and checks, make tests work for provider addresses.
