+++ tox.ini
@@ -12,7 +12,7 @@ setenv =
     PYTHONPATH = {toxinidir}
     CONFIG_FILE = {toxinidir}/config_local.ini
     AUTH_TOKEN_EXPIRATION=3153600000
-    PROVIDER_PRIVATE_KEY=0xc594c6e5def4bab63ac29eed19a134c130388f74f019bc74b8f4389df2837a58
+    PROVIDER_PRIVATE_KEY=0xfd5c1ccea015b6d663618850824154a3b3fb2882c46cefb05b9a93fea8c3d215
     TEST_PRIVATE_KEY1=0xef4b441145c1d0f3b4bc6d61d29f5c6e502359481152f869247c7a4244d45209
     TEST_PRIVATE_KEY2=0x5d75837394b078ce97bc289fa8d75e21000573520bfa7784a9d28ccaae602bf8
     OPERATOR_SERVICE_URL=https://operator-api.operator.dev-ocean.com
